Fixed segmentation fault bug
Also found some memory leaks, but I was able to fix them by changing the deleteMessage method.
On line 82 for the If() statement I previously had curr node as the value but curr node was pointing to a null pointer.  I changed it to previous node which fixed the memory leaks but my output is still not correct.
My first guess and what I am gonna test first is the delete method I believe it is deleting the wrong node.
My Second guess is I am reading faulty input from the file but this might be the least likely to be causing the bad output.